+++ packages/terra-application-links/CHANGELOG.md
@@ -4,6 +4,14 @@ Changelog
 Unreleased
 ----------
 
+6.6.0 - (July 16, 2019)
+------------------
+### Added
+* Added documentation on required peerDependencies
+
+### Changed
+* Updated to interface with react-intl's Intl context rather than old React Context API
+
 6.5.0 - (July 10, 2019)
 ------------------
 ### Changed